feature/SP-109: full FIS sync: fixed sorting of new employees.
